+++ test/SemanticObjectAnonymousSub.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace VirtualAssembly\Semantizer;
+
+class SemanticObjectAnonymousSub extends SemanticObjectAnonymous {
+
+    public function __construct(Semantizer $semantizer) {
+        parent::__construct($semantizer, "foaf:testType");
+    }
+
+}